added front end for the smart contract

+++ lottery/contracts/Lottery.sol
@@ -3,32 +3,35 @@ pragma solidity ^0.4.17;
 contract Lottery {
     address public manager;
     address[] public players;
-    
+
     function Lottery() public {
         manager = msg.sender;
     }
-    
+
     function enter() public payable {
         require(msg.value > .01 ether);
         players.push(msg.sender);
     }
-    
-    function random() private view returns (uint) {
-        return uint(keccak256(block.difficulty, now, players));
+
+    function random() private view returns (uint256) {
+        return uint256(keccak256(block.difficulty, now, players));
     }
-    
+
     function pickWinner() public restricted {
-        uint index = random() % players.length;
+        uint256 index = random() % players.length;
         players[index].transfer(this.balance);
+        // address lastWinner = players[index];
         players = new address[](0);
+
+        // return lastWinner;
     }
-    
+
     modifier restricted() {
         require(msg.sender == manager);
         _;
     }
-    
+
     function getPlayers() public view returns (address[]) {
         return players;
     }
-}   
+}
